Jump back to my loader after the loaded program is terminated
(Incomplete) Succeed to find __exit_funcs

(incomplete) Reference indirectly __exit_funcs

The value of __exit_funcs is 0x6b9100 and it is not the value it points.
so, void** addr = 0x6b9110; __exit_funcs = *addr is a correct way to handle.

Jump back to my loader

It works well, but the mangling code is hard-coded, so I need some more research.